refactor: extract config
