+++ packages/blocks/src/bookmark-block/components/bookmark-operation-popper.ts
@@ -1,95 +1,10 @@
 import { WithDisposable } from '@blocksuite/lit';
 import type { BaseBlockModel } from '@blocksuite/store';
-import { css, html, LitElement, nothing, type TemplateResult } from 'lit';
+import { css, html, LitElement, nothing } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import { repeat } from 'lit/directives/repeat.js';
 
-import { toast } from '../..//components/toast.js';
-import { copyBlocks } from '../../__internal__/clipboard/utils/commons.js';
-import { getBlockElementByModel } from '../../__internal__/index.js';
-import {
-  CopyIcon,
-  DeleteIcon,
-  DuplicateIcon,
-  RefreshIcon,
-} from '../../icons/index.js';
-import type { BookmarkBlockComponent } from '../bookmark-block.js';
-import type { BookmarkBlockModel } from '../bookmark-model.js';
-import {
-  cloneBookmarkProperties,
-  reloadBookmarkBlock,
-  tryGetBookmarkAPI,
-} from '../utils.js';
-
-export type MenuActionCallback = (type: Operation['type']) => void;
-
-type Operation = {
-  type: 'reload' | 'copy' | 'delete' | 'duplicate';
-  icon: TemplateResult;
-  label: string;
-  action: (
-    model: BaseBlockModel<BookmarkBlockModel>,
-    callback?: MenuActionCallback,
-    element?: HTMLElement
-  ) => void;
-  showWhen?: (model: BaseBlockModel<BookmarkBlockModel>) => boolean;
-  disableWhen?: (model: BaseBlockModel<BookmarkBlockModel>) => boolean;
-  divider?: boolean;
-};
-const operations: Operation[] = [
-  {
-    type: 'copy',
-    icon: CopyIcon,
-    label: 'Copy',
-    action: async (model, callback) => {
-      await copyBlocks([model]);
-      toast('Copied link to clipboard');
-      callback?.('copy');
-    },
-  },
-  {
-    type: 'duplicate',
-    icon: DuplicateIcon,
-    label: 'Duplicate',
-    disableWhen: model => model.page.readonly,
-    action: (model, callback) => {
-      const { page } = model;
-
-      const parent = page.getParent(model);
-      const index = parent?.children.indexOf(model);
-
-      const clonedProps = cloneBookmarkProperties(model);
-
-      page.addBlock('affine:bookmark', clonedProps, parent, index);
-
-      callback?.('duplicate');
-    },
-  },
-  {
-    type: 'reload',
-    icon: RefreshIcon,
-    label: 'Reload',
-    disableWhen: model => model.page.readonly || !tryGetBookmarkAPI(),
-    action: (model, callback) => {
-      reloadBookmarkBlock(
-        model,
-        getBlockElementByModel(model) as BookmarkBlockComponent,
-        true
-      );
-      callback?.('reload');
-    },
-  },
-  {
-    type: 'delete',
-    icon: DeleteIcon,
-    label: 'Delete',
-    disableWhen: model => model.page.readonly,
-    action: (model, callback) => {
-      model.page.deleteBlock(model);
-      callback?.('delete');
-    },
-  },
-];
+import { type MenuActionCallback, moreOperations } from './config.js';
 
 @customElement('bookmark-operation-menu')
 export class BookmarkOperationMenu extends WithDisposable(LitElement) {
@@ -132,7 +47,9 @@ export class BookmarkOperationMenu extends WithDisposable(LitElement) {
 
   override render() {
     const menuItems = repeat(
-      operations.filter(({ showWhen = () => true }) => showWhen(this.model)),
+      moreOperations.filter(({ showWhen = () => true }) =>
+        showWhen(this.model)
+      ),
       ({ type }) => type,
       ({ type, icon, label, action, divider, disableWhen = () => false }) => {
         return html`<icon-button
